data decoding feature added to test canister
