+++ src/subscription_manager.rs
@@ -194,6 +194,7 @@ async fn distribute_event(event: Event) {
                     timestamp: current_timestamp(),
                     namespace: event.namespace.clone(),
                     data: event.data.clone(),
+                    tx_hash: event.tx_hash.clone(),
                     headers: event.headers.clone(),
                     source: ic_cdk::api::id(),
                     filter: None, // We don't need to store the filter in the notification
