+++ src/test_canister1/Cargo.toml
@@ -11,6 +11,7 @@ serde_json = "1.0"
 evm_logs_types = { path = "../../evm_logs_types" } 
 ic-cdk-macros = "0.9"
 ic-web3-rs = "0.1.9"
+hex = "0.4.3"
 
 [lib]
 crate-type = ["cdylib"]
