+++ src/test_canister1/src/lib.rs
@@ -1,82 +1,17 @@
 use ic_cdk::api::call::call;
-use candid::{Principal};
 use ic_cdk_macros::{update, query, init};
+use candid::Principal;
 use std::cell::RefCell;
-use evm_logs_types::{SubscriptionRegistration, RegisterSubscriptionResult, EventNotification, UnsubscribeResult, ICRC16Value};
-use candid::{CandidType, Deserialize};
-use ic_web3_rs::ethabi::{decode, ParamType, Token};
-use ic_web3_rs::types::{U256, H160};
-use std::convert::{TryFrom, TryInto};
-use std::str::FromStr;
+use evm_logs_types::{SubscriptionRegistration, RegisterSubscriptionResult, EventNotification, UnsubscribeResult};
+use ic_web3_rs::ethabi::{decode, ParamType};
+use std::convert::TryInto;
+use hex;
 
-#[derive(CandidType, Deserialize, Clone, Debug)]
-pub struct SwapEventData {
-    pub sender: SolidityToken,
-    pub amount0: SolidityToken,
-    pub amount1: SolidityToken,
-}
+mod read_contract;
 
-/// SolidityToken is a representation of a web3_rs Token, but with CandidType support
-#[derive(CandidType, Deserialize, Clone, Debug)]
-pub enum SolidityToken {
-    Address(String),
-    FixedBytes(ic_web3_rs::ethabi::FixedBytes),
-    Bytes(ic_web3_rs::ethabi::Bytes),
-    Int(String),
-    Uint(String),
-    Bool(bool),
-    String(String),
-    FixedArray(Vec<SolidityToken>),
-    Array(Vec<SolidityToken>),
-    Tuple(Vec<SolidityToken>),
-}
+use read_contract::{SolidityToken, SwapEventData};
 
-impl From<Token> for SolidityToken {
-    fn from(token: Token) -> Self {
-        match token {
-            Token::Address(address) => SolidityToken::Address(format!("{:?}", address)),
-            Token::FixedBytes(bytes) => SolidityToken::FixedBytes(bytes),
-            Token::Bytes(bytes) => SolidityToken::Bytes(bytes),
-            Token::Int(int) => SolidityToken::Int(format!("{}", int)),
-            Token::Uint(uint) => SolidityToken::Uint(format!("{}", uint)),
-            Token::Bool(boolean) => SolidityToken::Bool(boolean),
-            Token::String(string) => SolidityToken::String(string),
-            Token::FixedArray(tokens) => {
-                SolidityToken::FixedArray(tokens.into_iter().map(SolidityToken::from).collect())
-            }
-            Token::Array(tokens) => {
-                SolidityToken::Array(tokens.into_iter().map(SolidityToken::from).collect())
-            }
-            Token::Tuple(tokens) => {
-                SolidityToken::Tuple(tokens.into_iter().map(SolidityToken::from).collect())
-            }
-        }
-    }
-}
-
-impl From<SolidityToken> for Token {
-    fn from(token: SolidityToken) -> Self {
-        match token {
-            SolidityToken::Address(address) => Token::Address(H160::from_str(&address).unwrap()),
-            SolidityToken::FixedBytes(bytes) => Token::FixedBytes(bytes),
-            SolidityToken::Bytes(bytes) => Token::Bytes(bytes),
-            SolidityToken::Int(int) => Token::Int(U256::from_str_radix(&int, 10).unwrap()),
-            SolidityToken::Uint(uint) => Token::Uint(U256::from_str_radix(&uint, 10).unwrap()),
-            SolidityToken::Bool(boolean) => Token::Bool(boolean),
-            SolidityToken::String(string) => Token::String(string),
-            SolidityToken::FixedArray(tokens) => {
-                Token::FixedArray(tokens.into_iter().map(Token::from).collect())
-            }
-            SolidityToken::Array(tokens) => {
-                Token::Array(tokens.into_iter().map(Token::from).collect())
-            }
-            SolidityToken::Tuple(tokens) => {
-                Token::Tuple(tokens.into_iter().map(Token::from).collect())
-            }
-        }
-    }
-}
-
+const DECIMALS: i128 = 10i128.pow(18); // Constant representing 10^18
 
 thread_local! {
     static NOTIFICATIONS: RefCell<Vec<EventNotification>> = RefCell::new(Vec::new());
@@ -173,15 +108,13 @@ async fn get_subscriptions(canister_id: Principal) -> Vec<evm_logs_types::Subscr
 
 fn decode_event_data(data: Vec<u8>) -> Result<Vec<SolidityToken>, String> {
     // index_topic_1 address sender, index_topic_2 address recipient, int256 amount0, int256 amount1, uint160 sqrtPriceX96, uint128 liquidity, int24 tick
-
+    
     let param_types = vec![
-        ParamType::Address, // sender
-        ParamType::Address, // recipient
         ParamType::Int(256), // amount0
         ParamType::Int(256), // amount1
-        ParamType::Uint(160), // sqrtPriceX96
-        ParamType::Uint(128), // liquidity
-        ParamType::Int(24),   // tick
+        ParamType::Int(256), // sqrtPriceX96
+        ParamType::Int(256), // liquidity
+        ParamType::Int(256),   // tick
     ];
 
     let decoded_tokens = decode(&param_types, &data)
@@ -191,19 +124,24 @@ fn decode_event_data(data: Vec<u8>) -> Result<Vec<SolidityToken>, String> {
     Ok(result)
 }
 
+
 #[query]
 fn get_swap_events_data() -> Vec<SwapEventData> {
     NOTIFICATIONS.with(|notifications| {
         let mut result = Vec::new();
         for notification in notifications.borrow().iter() {
-            let data: Vec<u8> = notification.data.clone().try_into().unwrap();
+            let data: Vec<u8> = hex::decode(String::try_from(notification.data.clone()).unwrap()).unwrap();
             let decoded_data = decode_event_data(data);
             match decoded_data {
                 Ok(tokens) => {
                     let event_data = SwapEventData {
-                        sender: SolidityToken::from(tokens[0].clone()),
-                        amount0: SolidityToken::from(tokens[2].clone()),
-                        amount1: SolidityToken::from(tokens[3].clone()),
+                        tx_hash: notification.tx_hash.clone(),
+                        amount0: tokens[0].clone(), 
+                        amount1: tokens[1].clone(), 
+                        sqrtPriceX96: tokens[2].clone(),
+                        liquidity: tokens[3].clone(),
+                        tick: tokens[4].clone(),
+
                     };
 
                     result.push(event_data);
